fastapi/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-escpos==3.1
pyusb==1.3.1
libusb-package==1.0.26.3
python-dotenv==1.0.1
